Modificado: /save agora retore um ResponseEntity
- Modificado: Price agora recebe double e Double
- Modificado:
Classes Volumes e Price de value para volume e price devido ser palavra reservada em SQL
- Adicionado: Construtores vazios para Price e Volumes

+++ backend/stockcontrolapi/src/main/java/com/cmoiss/stockcontrolapi/dtos/request/VolumeVariationDTO.java
@@ -11,6 +11,6 @@ import java.math.BigDecimal;
 @Setter
 public class VolumeVariationDTO {
     private Double volume;  // Valor direto (não objeto Volumes)
-    private BigDecimal price; // Valor direto (não objeto Price)
+    private Double price; // Valor direto (não objeto Price)
     private Integer quantity;
 }
